backend/requirements.txt
Flask
requests
python-dotenv
google-generativeai
scikit-learn
sarvamai
Flask-Cors